# allow frameworks to specify which schemes Carthage should build