By specifying an optional `.carthage_schemes` file at the root of a framework repo, Carthage can skip much of the work involved in evaluating which schemes to build in a directory.

If present, the `.carthage_schemes` file should be a newline-separated list of scheme names that Carthage should build.

The work involved in evaluating which schemes to build can be substantial due to requiring `xcodebuild` invocations. In the case of the Realm frameworks, for example, using this approach shaves minutes off the total update time.

Since frameworks are sometimes useful for purposes internal to a repo and not meant to be consumed externally, specifying a subset of schemes to build in `.carthage_schemes` can also great speed up build times.

This can also be useful to work around certain Xcode bugs, such as when Xcode considers unit test bundle products to actually be framework products [rdar://29407087](http://www.openradar.me/29407087).