@jpsim 2016-11-29
Building both `Realm.framework` and `RealmSwift.framework` is ~40% faster with this PR (499s vs 839s), since it works around [rdar://29407087](http://www.openradar.me/29407087).

---

Current `master`:

```shell
$ time carthage update --platform osx
*** Fetching realm-cocoa
*** Checking out realm-cocoa at "a772de1626c9e6b3279c508375a539c29c77244c"
*** xcodebuild output can be found in /var/folders/np/5y4sgt8x3r13sk7x0n31z_3m0000gn/T/carthage-xcodebuild.JBCzD7.log
*** Building scheme "Realm" in Realm.xcworkspace
*** Building scheme "Object Server Tests" in Realm.xcworkspace
*** Building scheme "RealmSwift" in Realm.xcworkspace
carthage update --platform osx  839.58s user 86.48s system 262% cpu 5:52.89 total
```

This PR:

```shell
$ time carthage update --platform osx
*** Fetching realm-cocoa
*** Checking out realm-cocoa at "a772de1626c9e6b3279c508375a539c29c77244c"
*** xcodebuild output can be found in /var/folders/np/5y4sgt8x3r13sk7x0n31z_3m0000gn/T/carthage-xcodebuild.VXjlfY.log
*** Building scheme "Realm" in Realm.xcworkspace
*** Building scheme "RealmSwift" in Realm.xcworkspace
carthage update --platform osx  499.41s user 45.25s system 238% cpu 3:48.07 total
```

---

Unfortunately, Carthage still builds `Realm.framework` once too many, since `RealmSwift.framework` uses it as a dependency, and Carthage always cleans before building, but that's a different battle.