@younata 2016-11-29
That looks like a pretty good idea, and has been something I've been wanting to do myself.

Would it be possible to add some test coverage for this?