@Pr0Ger 2016-11-29
I think it's better to give this kind of control to project instead of framework. For example project can build only 'Realm' scheme, or only 'RealmSwift' dependes on what it is required.

Relevant issue: #1227