@jpsim 2016-11-29
> I think it's better to give this kind of control to project instead of framework.

That's a nice idea, but orthogonal to this feature. It doesn't address preventing internal-only frameworks from being exposed externally.

For example, Realm uses a framework called `PlaygroundFrameworkWrapper` to enable using a prebuilt framework in a playground (typically only allowed for framework _targets_ defined in an Xcode workspace), which is only used for packaging purposes, and should never be exposed externally. We've had to unshare this framework scheme to cater to Carthage's behavior (realm/realm-cocoa@9c38192), which isn't ideal and shouldn't be required of any project maintainer since it adversely affects non-Carthage use cases.