@jpsim 2016-11-29
> Would it be possible to add some test coverage for this?

I'm happy to add tests and documentation, but before I invest time into that, I'd like to get a sense from Carthage maintainers whether or not this is an idea they're generally comfortable with.