@mdiep 2016-12-01
Thanks for the pull request. I always really appreciate when people open a PR instead of expecting maintainers to do all the work. ☺️ 

I do feel a little uncomfortable with this because it opens the door to a lot of additional configuration and complexity—and Carthage has eschewed all configuration. I feel this both on an implementation level (where the codebase needs to be refactored to handle the existing complexity) and conception level (where we'd probably want a larger vision for what configuration we want to allow and how it works).

But I also respect that this is a very pragmatic choice. That's part of why I've let this sit for a couple days before responding. Idealism vs. pragmatism is a hard line to manage.

I noticed that your radar had a response. Have you tried out the solution they gave? Knowing whether that solves the issue at hand seems like the next step.