@jpsim 2016-12-01
> I always really appreciate when people open a PR instead of expecting maintainers to do all the work.

Same! And this is a better way to spawn a more concrete conversation than by having a hand-wavy feature request issue.

> I do feel a little uncomfortable with this because it opens the door to a lot of additional configuration and complexity—and Carthage has eschewed all configuration. I feel this both on an implementation level (where the codebase needs to be refactored to handle the existing complexity) and conception level (where we'd probably want a larger vision for what configuration we want to allow and how it works).

I can appreciate that. However in practice, as a library maintainer with a fairly complex repo structure, I can safely say that conforming to Carthage's expectations has been more of a burden, and needed more refactoring than this simple configuration option involves.

For example, there's no way currently to have internal-only frameworks (like the `PlaygroundFrameworkWrapper` example above) without also making compromises, which in some cases aren't viable, such as unsharing the framework scheme in that specific example.

If we explicitly controlled which schemes we wanted Carthage to consider building, there would be less of a burden on library maintainers to consider the Carthage implications of adding another framework in their repo, sometimes deeply nested in an `examples/` subdirectory.

Please keep in mind as well that this file is entirely optional and won't affect the majority of library maintainers with very simple project structures.

As for wanting to explore a more general strategy for configuration, that's a fair argument and I respect if that's the outcome of this. However, it is disappointing. I'd much rather see this implemented with a warning saying that it's a temporary solution until a more comprehensive configuration approach can be designed and implemented. As you can tell, it's also not a feature that will require a high maintenance burden (it's dead simple for this reason), so it could conceivably be preserved even after a different configuration strategy is introduced.

> I noticed that your radar had a response. Have you tried out the solution they gave? Knowing whether that solves the issue at hand seems like the next step.

Yes, but that's only a partial solution to one of the motivating factors for this feature (that Xcode behavior). It doesn't address the "internal-only framework" scenario presented above.