@guidomb 2017-01-04
This could be a solution to the problem of working with dependencies like https://github.com/aws/aws-sdk-ios. They have lots of scheme and the entire project takes around 30 minutes to compile. Even if as a library consumer I only want to use AWS S3 I need to be build all their services frameworks. Users of AWS iOS SDK that include it via CocoaPods don't have this issue because they can use subspec but this is not an option with Carthage.

Although we can discuss if the way AWS SDK is structure is adequate or not, having the solution that @jpsim propose will make our life easier and will lower the barrier for library maintainers to adopt Carthage.