@guidomb 2017-01-04
@jpsim Is this an option only for the library maintainer or for libraries consumers too? Because if this is only for maintainers then what I said before about AWS SDK does not apply and maybe a solution like #1227 would be better.