@jpsim 2017-01-04
As-is, this would only be for library maintainers.