@mdiep 2017-01-05
> Since frameworks are sometimes useful for purposes internal to a repo and not meant to be consumed externally

Do you have any examples of this? Internal frameworks that are part of shared schemes? (The `Object Server Tests` case above doesn't fit since it was an Xcode/configuration issue.)

_(Sorry for the delay. But now that ReactiveSwift 1.0 has shipped, I should be able to wrap up this discussion.)_