@jpsim 2017-01-05
An example is the `PlaygroundFrameworkWrapper` that I mentioned in a previous issue.